+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Add `Ecto.Query.API.splice/1` support https://github.com/plausible/ecto_ch/pull/240
-- Add `Ecto.Query.API.constant/1` support
+- Add `Ecto.Query.API.constant/1` support https://github.com/plausible/ecto_ch/pull/243
 
 ## 0.8.1 (2025-08-20)